Add files via upload
Proof of concept of the sessions stealing exploit that i unintentionally discovered and gave to the admins of macrochasm.net. I first assumed that it was something wrong with the macrochasm plugins but then came to an understanding that it was a problem with Mojangs Legacy Auth. Just to clear things up, unlike others have said I did not use an offline mode game it was online. The client i used was DoomsDay b2 which was released 2 years ago which is a skid of Darkmagician6 as far as I've been told. The feature to login with an alt was also added in release b2 and the devs were known to have used the exploit. Accounts i logged into in order are NOOP, STALIN, 0x22, Notch, Antvenom, popbob, TheCampingRusher and finally jared2013.